Create RecordShowRightDrawerOpenRecordButton and implement shortcut
