+++ packages/twenty-front/src/modules/ui/layout/show-page/components/ShowPageSubContainer.tsx
@@ -1,7 +1,6 @@
 import { RecordShowRightDrawerActionMenu } from '@/action-menu/components/RecordShowRightDrawerActionMenu';
+import { RecordShowRightDrawerOpenRecordButton } from '@/action-menu/components/RecordShowRightDrawerOpenRecordButton';
 import { ActivityTargetableObject } from '@/activities/types/ActivityTargetableEntity';
-import { useCommandMenu } from '@/command-menu/hooks/useCommandMenu';
-import { getLinkToShowPage } from '@/object-metadata/utils/getLinkToShowPage';
 import { isNewViewableRecordLoadingState } from '@/object-record/record-right-drawer/states/isNewViewableRecordLoading';
 import { CardComponents } from '@/object-record/record-show/components/CardComponents';
 import { FieldsCard } from '@/object-record/record-show/components/FieldsCard';
@@ -16,9 +15,7 @@ import { SingleTabProps, TabList } from '@/ui/layout/tab/components/TabList';
 import { useTabList } from '@/ui/layout/tab/hooks/useTabList';
 import { useIsMobile } from '@/ui/utilities/responsive/hooks/useIsMobile';
 import styled from '@emotion/styled';
-import { Link } from 'react-router-dom';
 import { useRecoilState, useRecoilValue } from 'recoil';
-import { Button, IconArrowsDiagonal, getOsControlSymbol } from 'twenty-ui';
 
 const StyledShowPageRightContainer = styled.div<{ isMobile: boolean }>`
   display: flex;
@@ -43,10 +40,6 @@ const StyledContentContainer = styled.div<{ isInRightDrawer: boolean }>`
 `;
 
 export const TAB_LIST_COMPONENT_ID = 'show-page-right-tab-list';
-
-const StyledLink = styled(Link)`
-  text-decoration: none;
-`;
 
 type ShowPageSubContainerProps = {
   layout?: RecordLayout;
@@ -119,8 +112,6 @@ export const ShowPageSubContainer = ({
   const displaySummaryAndFields =
     layout && !layout.hideSummaryAndFields && !isMobile && !isInRightDrawer;
 
-  const { closeCommandMenu } = useCommandMenu();
-
   return (
     <>
       {displaySummaryAndFields && (
@@ -147,22 +138,10 @@ export const ShowPageSubContainer = ({
           <RightDrawerFooter
             actions={[
               <RecordShowRightDrawerActionMenu />,
-              <StyledLink
-                to={getLinkToShowPage(
-                  targetableObject.targetObjectNameSingular,
-                  recordFromStore,
-                )}
-                onClick={closeCommandMenu}
-              >
-                <Button
-                  title="Open"
-                  variant="primary"
-                  accent="blue"
-                  size="medium"
-                  Icon={IconArrowsDiagonal}
-                  hotkeys={[getOsControlSymbol(), '⏎']}
-                />
-              </StyledLink>,
+              <RecordShowRightDrawerOpenRecordButton
+                objectNameSingular={targetableObject.targetObjectNameSingular}
+                record={recordFromStore}
+              />,
             ]}
           />
         )}
